# Optimisations on `triangle_intersection`