Building on #17, this should bring some optimisations
```
ray_mesh_intersection/100_vertices
                        time:   [2.5537 us 2.5698 us 2.5862 us]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe
ray_mesh_intersection/10000_vertices
                        time:   [302.76 us 304.54 us 306.30 us]
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe
ray_mesh_intersection/1000000_vertices
                        time:   [35.273 ms 35.525 ms 35.841 ms]
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe

ray_mesh_intersection_no_intersection/100_vertices
                        time:   [2.4214 us 2.4359 us 2.4505 us]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
ray_mesh_intersection_no_intersection/10000_vertices
                        time:   [291.88 us 293.84 us 295.84 us]
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
ray_mesh_intersection_no_intersection/1000000_vertices
                        time:   [35.197 ms 35.413 ms 35.638 ms]
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe
```

* First commit: stop on first vertex that is closer instead of checking all 3
* Second commit: introduce a trait `TriangleTrait` instead of changing to the `Triangle` struct. The struct implies a copy while the trait doesn't. This probably could be done cleanlier but I was aiming for the smallest change