@aevyrie 2021-06-19
@mockersf this is great, some really significant gains! 

Can you help me understand the copy the second commit resolves?

`[Vec3; 3] -> Copy Vec3s into the triangle struct? -> struct Triangle -> v0(), v1(), v2()`

vs.

`[Vec3; 3] -> index directly into the Vec3 array -> v0(), v1(), v2()`

I assumed that the triangle struct could be optimized away at compile time, does this have something to do with the struct fields forcing a copy? Perhaps if the `Triangle` struct was a thin wrapper over `data: [Vec3; 3]` it would prevent this copy?