@mockersf 2021-06-19
> @mockersf this is great, some really significant gains!

Don't hesitate to run the benchmark on your machine, mine is very noisy, I can get +/- 5% between reruns on the same code

> Can you help me understand the copy the second commit resolves?

I'm not completely sure, it's just. something I tried and noticed an improvement 😄 

> I assumed that the triangle struct could be optimized away at compile time, does this have something to do with the struct fields forcing a copy? Perhaps if the `Triangle` struct was a thin wrapper over `data: [Vec3; 3]` it would prevent this copy?

Yup I tried one possible way to keep pointing at the same data that I felt was the simpler, but it could be done in other ways